+++ main.py
@@ -2,132 +2,36 @@ import pygame
 import os
 import time 
 import random
-from objects import Laser
-from objects import collide
+
+from objects.laser import Laser, collide
+from objects.player import Player
+from objects.enemy import Enemy
+
+class Game():
+    
+    def __init__(self, run=True, FPS=60, level=0, lives=0):
+        self.run = run
+        self.FPS = FPS
+        self.level = level
+        self.lives = lives
+
 
 pygame.font.init()
 
 WIDTH, HEIGHT = 500, 500
 WIN = pygame.display.set_mode((WIDTH, HEIGHT))
 pygame.display.set_caption("Space Shooter")
-
-RED_SPACESHIP = pygame.image.load(os.path.join("assets", "pixel_ship_red_small.png"))
-GREEN_SPACESHIP = pygame.image.load(os.path.join("assets", "pixel_ship_green_small.png"))
-BLUE_SPACESHIP = pygame.image.load(os.path.join("assets", "pixel_ship_blue_small.png"))
-YELLOW_SPACESHIP = pygame.image.load(os.path.join("assets", "pixel_ship_yellow.png"))
-
-RED_LASER = pygame.image.load(os.path.join("assets", "pixel_laser_red.png"))
-BLUE_LASER = pygame.image.load(os.path.join("assets", "pixel_laser_blue.png"))
-GREEN_LASER = pygame.image.load(os.path.join("assets", "pixel_laser_green.png"))
-YELLOW_LASER = pygame.image.load(os.path.join("assets", "pixel_laser_yellow.png"))
-
 BG = pygame.transform.scale(pygame.image.load(os.path.join("assets", "background-black.png")), (WIDTH, HEIGHT))
 
-class Ship:
-
-    COOLDOWN = 30
-    def __init__(self, x, y, health=100):
-        self.x = x
-        self.y = y
-        self.health = health
-        self.ship_img = None
-        self.laser_img = None
-        self.lasers = []
-        self.cool_down_counter = 0
+def play_game():
     
-    def get_width(self):
-        return self.ship_img.get_width()
-    
-    def get_height(self):
-        return self.ship_img.get_height()
+    game = Game() # Attention, redondance des objets de Game à gérer.
 
-    def draw(self, window):
-        window.blit(self.ship_img, (self.x, self.y))
-        for laser in self.lasers:
-            laser.draw(window)
-
-    def shoot(self):
-        if self.cool_down_counter == 0:
-            laser = Laser(self.x, self.y, self.laser_img)
-            self.lasers.append(laser)
-            self.cool_down_counter = 1
-
-    def cooldown(self):
-        if self.cool_down_counter >= self.COOLDOWN:
-            self.cool_down_counter = 0
-        elif self.cool_down_counter > 0:
-            self.cool_down_counter += 1
-
-    def move_lasers(self, vel, obj):
-        self.cooldown()
-        for laser in self.lasers:
-            laser.move(vel)
-            if laser.off_screen(HEIGHT):
-                self.lasers.remove(laser)
-            elif laser.collision(obj):
-                obj.health -= 10
-                self.lasers.remove(laser)
-
-
-class Player(Ship):
-    def __init__(self, x, y, health=100):
-        super().__init__(x, y, health)
-        self.ship_img = YELLOW_SPACESHIP
-        self.laser_img = YELLOW_LASER
-        self.mask = pygame.mask.from_surface(self.ship_img)
-        self.max_health = health
-
-    def move_lasers(self, vel, objs):
-        self.cooldown()
-        for laser in self.lasers:
-            laser.move(vel)
-            if laser.off_screen(HEIGHT):
-                self.lasers.remove(laser)
-            else:
-                for obj in objs:
-                    if laser.collision(obj):
-                        objs.remove(obj)
-                        self.lasers.remove(laser)
-    
-    def draw(self, window):
-        super().draw(window)
-        self.healthbar(window)
-
-    def healthbar(self, window):
-        pygame.draw.rect(window, (255, 0, 0), (self.x, self.y + self.ship_img.get_height() + 10, self.ship_img.get_width(), 10))
-        pygame.draw.rect(window, (0, 255, 0), (self.x, self.y + self.ship_img.get_height() + 10, self.ship_img.get_width() * (self.health/self.max_health), 10))
-
-
-class Enemy(Ship):
-
-    COLOR_MAP = {
-        "red": (RED_SPACESHIP, RED_LASER),
-        "green": (GREEN_SPACESHIP, GREEN_LASER),
-        "blue": (BLUE_SPACESHIP, BLUE_LASER),
-    }
-
-    def __init__(self, x, y, color, health=100):
-        super().__init__(x, y, health)
-        self.ship_img, self.laser_img = self.COLOR_MAP[color]
-        self.mask = pygame.mask.from_surface(self.ship_img)
-
-    def move(self, vel):
-        self.y += vel
-
-    def shoot(self):
-        if self.cool_down_counter == 0:
-            laser = Laser(self.x - 20, self.y, self.laser_img)
-            self.lasers.append(laser)
-            self.cool_down_counter = 1
-
-
-def main():
-    run = True
-    FPS = 60
-    level = 0
-    lives = 5
     main_font = pygame.font.SysFont("comicsans", 30)
-    lost_font = pygame.font.SysFont("comicsans", 60)
+
+    FPS=60
+    lives = 5
+    level = 0
 
     enemies = []
     wave_length = 5
@@ -141,38 +45,39 @@ def main():
 
     lost = False
     lost_count = 0
-
-    def redraw_window():
-        WIN.blit(BG, (0,0))
-        param_font_color = (255, 255, 255)
-        lives_label = main_font.render(f"Lives : {lives}", 1, param_font_color)
-        level_label = main_font.render(f"Level : {level}", 1, param_font_color)
+    # Defaut args : redraw_window(window=WIN, background=BG, li=lives, le=level, param_font_color=(255, 255, 255), enemies_list, player, lost):
+    def redraw_window(window, background, li, le, param_font_color, enemies_list, player, lost):
         
-        WIN.blit(lives_label, (10, 10))
-        WIN.blit(level_label, (WIDTH - level_label.get_width() - 10, 10))
+        window.blit(background, (0,0))
+        lives_label = main_font.render(f"Lives : {li}", 1, param_font_color)
+        level_label = main_font.render(f"Level : {le}", 1, param_font_color)
+        window.blit(lives_label, (10, 10))
+        window.blit(level_label, (window.get_width() - level_label.get_width() - 10, 10))
         
-        for enemy in enemies:
-            enemy.draw(WIN)
+        for enemy in enemies_list:
+            enemy.draw(window)
         
-        player.draw(WIN)
+        player.draw(window)
 
         if lost:
+            lost_font = pygame.font.SysFont("comicsans", 60)
             lost_label = lost_font.render("GAME OVER !", 1, param_font_color)
-            WIN.blit(lost_label,(int(WIDTH/2 - lost_label.get_width()/2), int(HEIGHT/2)))
+            
+            window.blit(lost_label,(int(window.get_width()/2 - lost_label.get_width()/2), int(window.get_height()/2)))
 
         pygame.display.update()
         
-    while run:
-        clock.tick(FPS)
-        redraw_window()
+    while game.run:
+        clock.tick(game.FPS)
+        redraw_window(window=WIN, background=BG, li=lives, le=level, param_font_color=(255, 255, 255), enemies_list=enemies, player=player, lost=lost)
 
         if lives <=0 or player.health <=0:
             lost = True
             lost_count += 1
 
         if lost:
-            if lost_count > FPS*5:
-                run = False
+            if lost_count > game.FPS*5:
+                game.run = False
             else:
                 continue 
 
@@ -183,7 +88,8 @@ def main():
                 enemy = Enemy(random.randrange(50, WIDTH-100), random.randrange(-1500, -100), random.choice(["red", "blue", "green"]))
                 enemies.append(enemy)
 
-        redraw_window()
+        redraw_window(window=WIN, background=BG, li=lives, le=level, param_font_color=(255, 255, 255), enemies_list=enemies, player=player, lost=lost)
+
 
         for event in pygame.event.get():
             if event.type == pygame.QUIT:
@@ -204,7 +110,7 @@ def main():
 
         for enemy in enemies[:]:
             enemy.move(enemy_vel)
-            enemy.move_lasers(laser_vel, player)
+            enemy.move_lasers(laser_vel, player, HEIGHT)
 
             if random.randrange(0, 2*60) == 1:
                 enemy.shoot()
@@ -217,7 +123,7 @@ def main():
                 lives -= 1
                 enemies.remove(enemy)
             
-        player.move_lasers(-laser_vel, enemies)
+        player.move_lasers(-laser_vel, enemies, HEIGHT)
 
 def main_menu():
     run = True
@@ -231,7 +137,7 @@ def main_menu():
             if event.type == pygame.QUIT:
                 run =False
             if event.type == pygame.MOUSEBUTTONDOWN:
-                main()
+                play_game()
     pygame.quit()
 
 main_menu()
