+++ objects/enemy.py
@@ -1,0 +1,27 @@
+import pygame
+from objects.ship import Ship
+from objects.laser import Laser
+from objects.img_manager import RED_SPACESHIP, RED_LASER, GREEN_SPACESHIP, GREEN_LASER, BLUE_SPACESHIP, BLUE_LASER
+
+class Enemy(Ship):
+
+    COLOR_MAP = {
+        "red": (RED_SPACESHIP, RED_LASER),
+        "green": (GREEN_SPACESHIP, GREEN_LASER),
+        "blue": (BLUE_SPACESHIP, BLUE_LASER),
+    }
+
+    def __init__(self, x, y, color, health=100):
+        super().__init__(x, y, health)
+        self.ship_img, self.laser_img = self.COLOR_MAP[color]
+        self.mask = pygame.mask.from_surface(self.ship_img)
+
+    def move(self, vel):
+        self.y += vel
+
+    def shoot(self):
+        if self.cool_down_counter == 0:
+            laser = Laser(self.x - 20, self.y, self.laser_img)
+            self.lasers.append(laser)
+            self.cool_down_counter = 1
+
