+++ objects/img_manager.py
@@ -1,0 +1,12 @@
+import pygame 
+from os.path import join
+
+RED_SPACESHIP = pygame.image.load(join("assets", "pixel_ship_red_small.png"))
+GREEN_SPACESHIP = pygame.image.load(join("assets", "pixel_ship_green_small.png"))
+BLUE_SPACESHIP = pygame.image.load(join("assets", "pixel_ship_blue_small.png"))
+YELLOW_SPACESHIP = pygame.image.load(join("assets", "pixel_ship_yellow.png"))
+
+RED_LASER = pygame.image.load(join("assets", "pixel_laser_red.png"))
+BLUE_LASER = pygame.image.load(join("assets", "pixel_laser_blue.png"))
+GREEN_LASER = pygame.image.load(join("assets", "pixel_laser_green.png"))
+YELLOW_LASER = pygame.image.load(join("assets", "pixel_laser_yellow.png"))
