+++ objects/laser.py
